@@ -1,51 +1,58 @@
+# frozen_string_literal: true
+
+# Module for formatting a record as Turtle
 module TurtleFormatter
-  def each_property( result, ignore_pattern = nil, &block )
-    props = result_to_ttl( result )[:properties]
-    props.reject {|pv| ignore_pattern && pv[:p] =~ ignore_pattern}.each( &block )
+  def each_property(result, ignore_pattern = nil, &block)
+    props = result_to_ttl(result)[:properties]
+    props.reject { |pv| ignore_pattern && pv[:p] =~ ignore_pattern }.each(&block)
   end
 
-  def result_to_ttl( result )
-    ttl_value = {properties: []}
+  def result_to_ttl(result) # rubocop:disable Metrics/MethodLength
+    ttl_value = { properties: [] }
 
     result.map do |property, value|
-      next if value.respond_to?(:size) && value.size == 0
-      if property == "@id"
-        ttl_value[:uri] = format_ttl_value( value )
+      next if value.respond_to?(:'empty?') && value.empty?
+      if property == '@id'
+        ttl_value[:uri] = format_ttl_value(value)
       else
-        ttl_value[:properties] << {p: format_ttl_value( property ),
-                                   v: format_ttl_value( value )}
+        ttl_value[:properties] << { p: format_ttl_value(property),
+                                    v: format_ttl_value(value) }
       end
     end
 
     ttl_value
   end
 
-  def format_ttl_value( v )
+  # rubocop:disable Metrics/MethodLength, Metrics/AbcSize
+  # rubocop:disable Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
+  def format_ttl_value(value)
     f =
-      if v == nil
-        ""
-      elsif v.is_a?( Array )
-        v.map {|v| format_ttl_value v} .join( ", " )
-      elsif v.is_a? Numeric
-        v.to_s
-      elsif v =~ /\Ahttp:\/\/.*/
-        "<#{v.to_s}>"
-      elsif v =~ /\A[[:word:]]+:.*/
-        v.to_s
-      elsif [false,true].include?( v )
-        v.to_s
-      elsif v["@id"]
-        "<#{v["@id"]}>"
-      elsif v["@value"] && v["@type"]
-        "\"#{v["@value"]}\"^^#{format_ttl_value( v["@type"] )}"
-      elsif v["@value"]
-        "\"#{v["@value"]}\""
-      elsif v.is_a? String
-        "\"#{v}\""
+      if value.nil?
+        ''
+      elsif value.is_a?(Array)
+        value.map { |v| format_ttl_value(v) } .join(', ')
+      elsif value.is_a? Numeric
+        value.to_s
+      elsif value.respond_to?(:'match?') && value.match?(%r{\Ahttp://.*})
+        "<#{value}>"
+      elsif value.respond_to?(:'match?') && value.match?(/\A[[:word:]]+:.*/)
+        value.to_s
+      elsif [false, true].include?(value)
+        value.to_s
+      elsif value['@id']
+        "<#{value['@id']}>"
+      elsif value['@value'] && value['@type']
+        "\"#{value['@value']}\"^^#{format_ttl_value(value['@type'])}"
+      elsif value['@value']
+        "\"#{value['@value']}\""
+      elsif value.is_a? String
+        "\"#{value}\""
       else
-        "\"#{v.to_s}\"^^<#{v.class.name}> # warning: default formatting rule (likely to be a bug)"
+        "\"#{value}\"^^<#{value.class.name}> # warning: default formatting rule (likely a bug)"
       end
 
     f.html_safe
   end
+  # rubocop:enable Metrics/MethodLength, Metrics/AbcSize
+  # rubocop:enable Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
 end
