@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 # Command object providing a service for driving the DsAPI
-# rubocop:disable Metrics/ClassLength
-class QueryCommand < DataService
+class QueryCommand < DataService # rubocop:disable Metrics/ClassLength
   include TurtleFormatter
 
   attr_reader :all_results, :search_results, :error_message
@@ -10,94 +9,94 @@ class QueryCommand < DataService
 
   ASPECTS = {
     saon:       SearchAspect.new(:saon,
-                                 "ppd:propertyAddress",
-                                 key_property: "lrcommon:saon",
-                                 aspect_key_property: "ppd:propertyAddressSaon",
-                                 presentation_label: "secondary name"),
+                                 'ppd:propertyAddress',
+                                 key_property: 'lrcommon:saon',
+                                 aspect_key_property: 'ppd:propertyAddressSaon',
+                                 presentation_label: 'secondary name'),
     paon:       SearchAspect.new(:paon,
-                                 "ppd:propertyAddress",
-                                 key_property: "lrcommon:paon",
-                                 aspect_key_property: "ppd:propertyAddressPaon",
-                                 presentation_label: "building name or no."),
+                                 'ppd:propertyAddress',
+                                 key_property: 'lrcommon:paon',
+                                 aspect_key_property: 'ppd:propertyAddressPaon',
+                                 presentation_label: 'building name or no.'),
     street:     SearchAspect.new(:street,
-                                 "ppd:propertyAddress",
-                                 key_property: "lrcommon:street",
-                                 aspect_key_property: "ppd:propertyAddressStreet"),
+                                 'ppd:propertyAddress',
+                                 key_property: 'lrcommon:street',
+                                 aspect_key_property: 'ppd:propertyAddressStreet'),
     town:       SearchAspect.new(:town,
-                                 "ppd:propertyAddress",
-                                 key_property: "lrcommon:town",
-                                 aspect_key_property: "ppd:propertyAddressTown"),
+                                 'ppd:propertyAddress',
+                                 key_property: 'lrcommon:town',
+                                 aspect_key_property: 'ppd:propertyAddressTown'),
     locality:   SearchAspect.new(:locality,
-                                 "ppd:propertyAddress",
-                                 key_property: "lrcommon:locality",
-                                 aspect_key_property: "ppd:propertyAddressLocality"),
+                                 'ppd:propertyAddress',
+                                 key_property: 'lrcommon:locality',
+                                 aspect_key_property: 'ppd:propertyAddressLocality'),
     district:   SearchAspect.new(:district,
-                                 "ppd:propertyAddress",
-                                 key_property: "lrcommon:district",
-                                 aspect_key_property: "ppd:propertyAddressDistrict"),
+                                 'ppd:propertyAddress',
+                                 key_property: 'lrcommon:district',
+                                 aspect_key_property: 'ppd:propertyAddressDistrict'),
     county:     SearchAspect.new(:county,
-                                 "ppd:propertyAddress",
-                                 key_property: "lrcommon:county",
-                                 aspect_key_property: "ppd:propertyAddressCounty"),
+                                 'ppd:propertyAddress',
+                                 key_property: 'lrcommon:county',
+                                 aspect_key_property: 'ppd:propertyAddressCounty'),
     postcode:   SearchAspect.new(:postcode,
-                                 "ppd:propertyAddress",
-                                 key_property: "lrcommon:postcode",
-                                 aspect_key_property: "ppd:propertyAddressPostcode"),
+                                 'ppd:propertyAddress',
+                                 key_property: 'lrcommon:postcode',
+                                 aspect_key_property: 'ppd:propertyAddressPostcode'),
     ptype:      FilterAspect.new(:ptype,
-                                 "ppd:propertyType",
-                                 values: %w(
+                                 'ppd:propertyType',
+                                 values: %w[
                                    lrcommon:detached
                                    lrcommon:semi-detached
                                    lrcommon:terraced
                                    lrcommon:flat-maisonette
                                    lrcommon:otherPropertyType
-                                 ),
+                                 ],
                                  uri_value: true,
-                                 presentation_label: "property type"),
+                                 presentation_label: 'property type'),
     nb:         FilterAspect.new(:nb,
-                                 "ppd:newBuild",
-                                 values: %w(true false),
+                                 'ppd:newBuild',
+                                 values: %w[true false],
                                  uri_value: false,
-                                 value_type: "xsd:boolean",
-                                 presentation_label: "new build?",
+                                 value_type: 'xsd:boolean',
+                                 presentation_label: 'new build?',
                                  value_labels: {
-                                   "true" => "new build only",
-                                   "false" => "existing buildings only"
+                                   'true' => 'new build only',
+                                   'false' => 'existing buildings only'
                                  }),
     et:         FilterAspect.new(:et,
-                                 "ppd:estateType",
-                                 values: %w(lrcommon:freehold lrcommon:leasehold),
+                                 'ppd:estateType',
+                                 values: %w[lrcommon:freehold lrcommon:leasehold],
                                  uri_value: true,
-                                 presentation_label: "estate type"),
+                                 presentation_label: 'estate type'),
     tc:         FilterAspect.new(:tc,
-                                 "ppd:transactionCategory",
-                                 values: %w(
+                                 'ppd:transactionCategory',
+                                 values: %w[
                                    ppd:standardPricePaidTransaction
                                    ppd:additionalPricePaidTransaction
-                                 ),
+                                 ],
                                  uri_value: true,
-                                 presentation_label: "transaction category"),
+                                 presentation_label: 'transaction category'),
     min_price:   RangeAspect.new(:min_price,
-                                 "ppd:pricePaid",
-                                 operator: "@ge",
-                                 presentation_label: "minimum price",
+                                 'ppd:pricePaid',
+                                 operator: '@ge',
+                                 presentation_label: 'minimum price',
                                  value_type: :numeric,
-                                 prompt: "%s is &pound;%s"),
+                                 prompt: '%s is &pound;%s'),
     max_price:   RangeAspect.new(:max_price,
-                                 "ppd:pricePaid",
-                                 operator: "@le",
-                                 presentation_label: "maximum price",
+                                 'ppd:pricePaid',
+                                 operator: '@le',
+                                 presentation_label: 'maximum price',
                                  value_type: :numeric,
-                                 prompt: "%s is &pound;%s"),
+                                 prompt: '%s is &pound;%s'),
     min_date:    RangeAspect.new(:min_date,
-                                 "ppd:transactionDate",
-                                 operator: "@ge",
-                                 presentation_label: "on or after",
+                                 'ppd:transactionDate',
+                                 operator: '@ge',
+                                 presentation_label: 'on or after',
                                  value_type: :date),
     max_date:    RangeAspect.new(:max_date,
-                                 "ppd:transactionDate",
-                                 operator: "@le",
-                                 presentation_label: "on or before",
+                                 'ppd:transactionDate',
+                                 operator: '@le',
+                                 presentation_label: 'on or before',
                                  value_type: :date)
   }.freeze
 
@@ -136,7 +135,7 @@ class QueryCommand < DataService
   end
 
   def save_results(ppd, query, options)
-    # Rails.logger.debug "About to ask DsAPI query: #{query.to_json}"
+    Rails.logger.debug "About to ask DsAPI query: #{query.to_json}"
     @all_results = ppd.query(query)
     @search_results = SearchResults.new(@all_results, options[:max])
   end
@@ -147,11 +146,15 @@ class QueryCommand < DataService
 
   def add_count_information(ppd, count_query)
     count_result = ppd.query(count_query)
-    count = count_result.first["@count"]
+    count = count_result.first['@count']
     @search_results.query_count("#{count}#{count == COUNT_LIMIT ? ' or more' : ''}")
   end
 
   def success?
     !error_message
+  end
+
+  def size
+    @search_results&.size
   end
 end
