@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 PpdExplorer::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
@@ -16,7 +18,7 @@ PpdExplorer::Application.configure do
 
   # Enable Rack::Cache to put a simple HTTP cache in front of your application
   # Add `rack-cache` to your Gemfile before enabling this.
-  # For large-scale production use, consider using a caching reverse proxy like nginx, varnish or squid.
+  # For large-scale production use, consider using a caching reverse proxy
   # config.action_dispatch.rack_cache = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this).
@@ -79,8 +81,8 @@ PpdExplorer::Application.configure do
   config.log_formatter = ::Logger::Formatter.new
 
   # Specify that we're not at the root
-  config.action_controller.relative_url_root = "/app/ppd"
-  config.relative_url_root = "/app/ppd"
+  config.action_controller.relative_url_root = '/app/ppd'
+  config.relative_url_root = '/app/ppd'
 end
 
 # JsRoutes.setup do |config|
