@@ -1,2 +1,4 @@
-Rails.application.config.assets.precompile += %w(hm_lr_logo.svg)
-Rails.application.config.assets.precompile += %w(hm_lr_logo.png)
+# frozen_string_literal: true
+
+Rails.application.config.assets.precompile += %w[hm_lr_logo.svg]
+Rails.application.config.assets.precompile += %w[hm_lr_logo.png]
