@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # Be sure to restart your server when you modify this file.
 
 # Add new inflection rules using the following format. Inflections
@@ -11,7 +13,7 @@
 # end
 
 ActiveSupport::Inflector.inflections(:en) do |inflect|
-  inflect.uncountable %w( PpdData PpdDataController )
+  inflect.uncountable %w[PpdData PpdDataController]
 end
 
 # These inflection rules are supported but not enabled by default:
